# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyticsAuthInput {
  password: string;
}

export interface AnalyticsAuthResult {
  ok: boolean;
}

export interface AnalyticsEventInput {
  sessionId: string;
  path: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /**
     * Seconds spent on the previous page
     * @nullable
     */
  duration?: number | null;
}

export interface TrackResult {
  ok: boolean;
}

export interface VisitorSession {
  id: number;
  sessionId: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  browser?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  isVpn: boolean;
  isProxy: boolean;
  pageViewCount: number;
  createdAt: string;
}

export interface StatEntry {
  label: string;
  count: number;
}

export interface AnalyticsSummary {
  totalVisitors: number;
  totalPageViews: number;
  todayVisitors: number;
  todayPageViews: number;
  weekVisitors: number;
  weekPageViews: number;
  topPages: StatEntry[];
  topCountries: StatEntry[];
  topReferrers: StatEntry[];
  topUtmSources: StatEntry[];
  deviceBreakdown: StatEntry[];
  browserBreakdown: StatEntry[];
  vpnCount: number;
  recentSessions: VisitorSession[];
}

export type ListAnalyticsSessionsParams = {
limit?: number;
offset?: number;
};