# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analytics
    description: Visitor analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analytics/auth:
    post:
      operationId: authenticateAnalytics
      tags: [analytics]
      summary: Validate analytics dashboard password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsAuthInput"
      responses:
        "200":
          description: Auth result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsAuthResult"

  /analytics/event:
    post:
      operationId: trackAnalyticsEvent
      tags: [analytics]
      summary: Track a page view event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventInput"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResult"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get aggregated analytics summary
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/sessions:
    get:
      operationId: listAnalyticsSessions
      tags: [analytics]
      summary: List recent visitor sessions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisitorSession"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AnalyticsAuthInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AnalyticsAuthResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    AnalyticsEventInput:
      type: object
      required: [sessionId, path]
      properties:
        sessionId:
          type: string
        path:
          type: string
        referrer:
          type: ["string", "null"]
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
          description: Seconds spent on the previous page

    TrackResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    VisitorSession:
      type: object
      required: [id, sessionId, isVpn, isProxy, pageViewCount, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        ipAddress:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
        browser:
          type: ["string", "null"]
        os:
          type: ["string", "null"]
        referrer:
          type: ["string", "null"]
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        isVpn:
          type: boolean
        isProxy:
          type: boolean
        pageViewCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    StatEntry:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer

    AnalyticsSummary:
      type: object
      required:
        - totalVisitors
        - totalPageViews
        - todayVisitors
        - todayPageViews
        - weekVisitors
        - weekPageViews
        - topPages
        - topCountries
        - topReferrers
        - topUtmSources
        - deviceBreakdown
        - browserBreakdown
        - vpnCount
        - recentSessions
      properties:
        totalVisitors:
          type: integer
        totalPageViews:
          type: integer
        todayVisitors:
          type: integer
        todayPageViews:
          type: integer
        weekVisitors:
          type: integer
        weekPageViews:
          type: integer
        topPages:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        topCountries:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        topReferrers:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        topUtmSources:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        deviceBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        browserBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatEntry"
        vpnCount:
          type: integer
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/VisitorSession"